+++ src/Package.php
@@ -4,20 +4,40 @@ namespace SymfonySi\LocalePack;
 
 class Package
 {
-    /** @var string Package name */
+    /**
+     * @var string Package name
+     */
     private $name;
 
-    /** @var string Package version */
+    /**
+     * @var string Package version
+     */
     private $version;
 
-    /** @var string Path to module configuration file */
+    /**
+     * @var string Path to module configuration file
+     */
     private $configFile;
 
-    /** @var string Path to generated package.xml file */
+    /**
+     * @var string Path to generated package.xml file
+     */
     private $packageFile;
 
-    /** @var array Output messages */
-    private $output = [];
+    /**
+     * @var string Package directory.
+     */
+    private $packageDir;
+
+    /**
+     * @var string Project directory.
+     */
+    private $rootDir;
+
+    /**
+     * @var array Output log messages
+     */
+    private $log = [];
 
     /**
      * Constructor.
@@ -47,26 +67,26 @@ class Package
     {
         // Patch config file first
         $this->patchConfigFile();
+
         // Patch package.xml file
-        $this->patchPackageFile();
+        $files = $this->patchPackageFile();
 
-        $tarFile = $this->rootDir.'/'.$this->name.'.tar';
-        $tgzFile = $this->rootDir.'/'.$this->name.'.tgz';
+        // Create tgz archive
+        $archiver = new Archiver($this->name.'-'.$this->version.'.tgz');
 
-        // Create tar archive
-        $phar = new \PharData($tarFile);
-        // add all files in the project
-        $phar->buildFromDirectory($this->packageDir);
+        // add all files defined in package.xml file
+        foreach($files as $file) {
+            $archiver->addFile($file, str_replace(realpath($this->packageDir), '', $file));
+        }
 
-        // Compress .tar file to .tgz
-        $phar->compress(\Phar::GZ, '.tgz');
-        rename($tgzFile, $this->rootDir.'/'.$this->name.'-'.$this->version.'.tgz');
-        $this->output[] = $this->name.'-'.$this->version.'.tgz created';
+        // add package.xml
+        $archiver->addFile(realpath($this->packageFile), str_replace(realpath($this->packageDir), '', realpath($this->packageFile)));
 
-        // Both files (.tar and .tgz) exist. Remove the temporary .tar archive
-        unlink($tarFile);
+        $archiver->compress();
 
-        return implode($this->output, "\n");
+        $this->log[] = $this->name.'-'.$this->version.'.tgz created';
+
+        return implode($this->log, "\n");
     }
 
     /**
@@ -79,45 +99,31 @@ class Package
         $dom->getElementsByTagName('version')->item(0)->nodeValue = $this->version;
         $dom->saveXML();
         $msg = $dom->save($this->configFile);
-        $this->output[] = 'Module config.xml updated: '.$msg.' bytes written.';
+        $this->log[] = 'Module config.xml updated: '.$msg.' bytes written.';
     }
 
     /**
      * Patches existing package.xml file with current time, date and md5 hashes of included files.
+     *
+     * @return array
      */
     private function patchPackageFile()
     {
+        $filesInPackage = [];
+
         $dom = new \DOMDocument;
         $dom->load($this->packageFile);
 
         $files = $dom->getElementsByTagName('file');
         foreach($files as $file) {
-            $filePath = $file->getAttribute('name');
-            $curNode = $file->parentNode;
-            while($curNode->nodeName == 'dir') {
-                $filePath = $curNode->getAttribute('name').'/'.$filePath;
-                $curNode = $curNode->parentNode;
-            }
+            $filePath = __DIR__.'/../package/'.$this->getFileFullPath($file);
 
-            $target = $curNode->getAttribute('name');
-            if ($target == 'magecommunity') {
-                $targetDir = 'app/code/community';
-            } else if ($target == 'magedesign') {
-                $targetDir = 'app/design';
-            } else if ($target == 'magelocale') {
-                $targetDir = 'app/locale';
-            } else if ($target == 'mageetc') {
-                $targetDir = 'app/etc';
-            } else if ($target == 'mageweb') {
-                $targetDir = '';
-            }
-
-            $filePath = __DIR__.'/../package/'.$targetDir.'/'.$filePath;
             if (file_exists($filePath)) {
                 $md5 = md5_file($filePath);
                 $file->setAttribute('hash', $md5);
+                $filesInPackage[] = realpath($filePath);
             } else {
-                $this->output[] = 'Warning: File '.basename($filePath).' not found.';
+                $this->log[] = 'Warning: File '.basename($filePath).' not found.';
             }
         }
 
@@ -130,6 +136,43 @@ class Package
 
         $dom->saveXML();
         $msg = $dom->save($this->packageFile);
-        $this->output[] = 'package.xml patched. '.$msg.' bytes written.';
+        $this->log[] = 'package.xml patched. '.$msg.' bytes written.';
+
+        return $filesInPackage;
+    }
+
+    /**
+     * Get file path for file from DoMNode.
+     *
+     * @param \DOMNode
+     *
+     * @return string
+     */
+    private function getFileFullPath(\DOMNode $file)
+    {
+        $curNode = $file->parentNode;
+        $filePath = $file->getAttribute('name');
+        while($curNode->nodeName == 'dir') {
+            $filePath = $curNode->getAttribute('name').'/'.$filePath;
+            $curNode = $curNode->parentNode;
+        }
+
+        $target = $curNode->getAttribute('name');
+        $targetDir = '';
+        if ($target == 'magecommunity') {
+            $targetDir = 'app/code/community';
+        } else if ($target == 'magedesign') {
+            $targetDir = 'app/design';
+        } else if ($target == 'magelocale') {
+            $targetDir = 'app/locale';
+        } else if ($target == 'mageetc') {
+            $targetDir = 'app/etc';
+        } else if ($target == 'mageweb') {
+            $targetDir = '';
+        }
+
+        $filePath = ($targetDir == '') ? $filePath : $targetDir.'/'.$filePath;
+
+        return $filePath;
     }
 }